changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11796 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Control/RemoteRunControl/ORRemoteRunModel.m
@@ -805,7 +805,7 @@ static NSString *ORRunRemoteConnectAtStart	= @"ORRunRemoteConnectAtStart";
     }
 }
 
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount
 {
     if(inNetSocket == socket){
         NSString* inString = [socket readString:NSASCIIStringEncoding];
